Refactor piano GUI types into submodule and clarify module structure
Moved Semitone, PointerId, KeySet, and ExternalKeySet from piano_gui.rs to piano_gui/types.rs as a proper submodule, following project guidelines. Updated all imports and removed the use of mod.rs. Also clarified the instructions file to explicitly allow submodules within module directories for better organization.
